Move lddmm scripts to scripts folder and rename strings to lddmm_strings
