+++ herbrain/lddmm_strings.py
